refactor: move method from `UserService` to `InvitationService`
